+++ docs/CHANGELOG.rst
@@ -5,8 +5,8 @@ Appendix C. Changelog
 :Author: Chris Warrick <chris@chriswarrick.com>
 :Copyright: © 2015, Chris Warrick.
 :License: BSD (see /LICENSE or :doc:`Appendix B <LICENSE>`.)
-:Date: 2015-07-06
-:Version: 0.1.1
+:Date: 2015-07-10
+:Version: 0.1.2
 
 .. index:: CHANGELOG
 
@@ -19,6 +19,9 @@ More information can be found on GitHub in the `releases section
 Version History
 ===============
 
+0.1.2
+    * Add mouse wheel support.
+
 0.1.1
     * Fix search form
     * Fix top-level passwords appearing twice
@@ -28,6 +31,9 @@ Version History
 
 0.1.0
     Initial release.
+
+0.1.2
+    * Add mouse wheel support.
 
 0.1.1
     * Fix search form